+++ src/Airbnb.Api/Controllers/ListingController.cs
@@ -1,4 +1,5 @@
 ﻿using Airbnb.AppService.Commands.Listing.CreateListing;
+using Airbnb.Core.Commands;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Airbnb.Api.Controllers;
@@ -7,6 +8,14 @@ namespace Airbnb.Api.Controllers;
 [Route("api/listings")]
 public class ListingController : ControllerBase
 {
+    private readonly ICommandDispatcher _commandDispatcher;
+
+    /// <inheritdoc />
+    public ListingController(ICommandDispatcher commandDispatcher)
+    {
+        _commandDispatcher = commandDispatcher;
+    }
+
     [HttpPost]
     public async Task<IActionResult> CreateListing([FromBody] CreateListingCommand command)
     {