+++ src/Airbnb.AppService/Commands/Listing/CreateListing/CreateListingCommandHandler.cs
@@ -1,6 +1,11 @@
-﻿namespace Airbnb.AppService.Commands.Listing.CreateListing;
+﻿using Airbnb.Core.Commands;
 
-public class CreateListingCommandHandler
+namespace Airbnb.AppService.Commands.Listing.CreateListing;
+
+public class CreateListingCommandHandler : ICommandHandler<CreateListingCommand>
 {
-    
+    public async Task HandleAsync(CreateListingCommand command)
+    {
+        throw new NotImplementedException();
+    }
 }