+++ src/Airbnb.AppService/Integrations/Listing/ListingCreated/ListingCreatedIntegrationEvent.cs
@@ -1,0 +1,8 @@
+﻿using Airbnb.SharedKernel.Events;
+
+namespace Airbnb.AppService.Integrations.Listing.ListingCreated;
+
+public sealed class ListingCreatedIntegrationEvent : IIntegrationEvent
+{
+    
+}