+++ src/Airbnb.AppService/Integrations/Listing/ListingCreated/ListingCreatedIntegrationHandler.cs
@@ -1,0 +1,11 @@
+﻿using Airbnb.Core.Events.IntegrationEvents;
+
+namespace Airbnb.AppService.Integrations.Listing.ListingCreated;
+
+public class ListingCreatedIntegrationHandler : IIntegrationEventHandler<ListingCreatedIntegrationEvent>
+{
+    public async Task HandleAsync(ListingCreatedIntegrationEvent @event)
+    {
+        throw new NotImplementedException();
+    }
+}