+++ src/Airbnb.Core/Commands/ICommand.cs
@@ -1,0 +1,9 @@
+﻿namespace Airbnb.Core.Commands;
+
+public interface ICommand
+{
+}
+
+public interface ICommand<out TResult> : ICommand
+{
+}