+++ src/Airbnb.Core/Commands/ICommandDispatcher.cs
@@ -1,0 +1,7 @@
+﻿namespace Airbnb.Core.Commands;
+
+public interface ICommandDispatcher
+{
+    Task<TResult> DispatchAsync<TResult>(ICommand<TResult> command);
+    Task<TResult> DispatchAsync<TCommand, TResult>(TCommand command) where TCommand : class, ICommand<TResult>;
+}