+++ src/Airbnb.Core/Commands/ICommandHandler.cs
@@ -1,0 +1,11 @@
+﻿namespace Airbnb.Core.Commands;
+
+public interface ICommandHandler<in TCommand, TResult> where TCommand : ICommand<TResult>
+{
+    Task<TResult> HandleAsync(TCommand command);
+}
+
+public interface ICommandHandler<in TCommand> where TCommand : ICommand
+{
+    Task HandleAsync(TCommand command);
+}