+++ src/Airbnb.Core/Events/DomainEvents/DomainEventDispatcher.cs
@@ -1,0 +1,6 @@
+﻿namespace Airbnb.Core.Events.DomainEvents;
+
+public class DomainEventDispatcher
+{
+    
+}