+++ src/Airbnb.Core/Events/IntegrationEvents/IIntegrationEventDispatcher.cs
@@ -1,0 +1,8 @@
+﻿using Airbnb.SharedKernel.Events;
+
+namespace Airbnb.Core.Events.IntegrationEvents;
+
+public interface IIntegrationEventDispatcher
+{
+    Task PublishAsync<TEvent>(TEvent @event) where TEvent : class, IIntegrationEvent;
+}